apps/api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.2
pydantic-settings==2.6.1
sqlalchemy==2.0.36
alembic==1.14.0
psycopg[binary]==3.2.3
geoalchemy2==0.15.2
shapely==2.0.6
pyproj==3.7.0
python-multipart==0.0.17
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
email-validator==2.2.0
redis==5.2.0
celery==5.4.0
httpx==0.28.1
pystac-client==0.8.5
planetary-computer==1.0.0
numpy==2.1.3
rasterio==1.4.2
boto3==1.35.66
pillow==11.0.0
structlog==24.4.0
prometheus-fastapi-instrumentator==7.0.0
python-dateutil==2.9.0.post0
pyogrio==0.10.0
orjson==3.10.12
pytest==8.3.4
pytest-asyncio==0.24.0
